+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 ### Fixes
 
 - [beta] Fix an issue that the editor unwontedly selects the nearest outline item when the Outline inspector is shown.
+- [beta] Fix an issue that the window subtitle did not update when the document is renamed.